This is a cleaned-up version of the GTFS Anomaly Detection that maps anomalies to a Folium map of the area, with toggle filters for each route and anomaly type, as well as route line maps.
This notebook runs on a CSV file of RT GTFS captured data, and the file will need to be manually recoded for whatever file is uploaded. 

The filters are not fully sorted, and could use some UI improvements, but everything seems to be working. 

There is some additional code optimization to be done, but this is a working version.