Newest version with better anomaly detection logic.
Still needs to have a widget created to more easily select filters on the final map. 

The anomalies placed on the map seem a bit chaotic, so I need to recheck the code, but I haven't worked with such a large timeframe before, so it's possible that there are some additional factors that create this apparent noise over time.